jobs: use parameters to set 'TAGS'
When submitting jobs to different LAVA instances, the jobs may fail if
the DUTs (Devices Under Test) do not have the required tags, such as
'lts' in the case of the db845c device.

To avoid this issue, use the '--parameters TAGS=tag1,tag2' option to
specify tags for jobs that must run on specific devices.

This ensures that users explicitly set TAGS=lts (or other required tags)
and allows job writers to target specific devices within a particular
group of DUTs.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>