templates: nfs: storage
Due to nvme drive being able to change, it has meant that its possible
for the wrong disk to be formatted. So instead, lets use the first
nvme disk, and then use the second partition after the ESP for the
scratch disk. Along with this changeset, I've simplied the existing
storage block.

Signed-off-by: Benjamin Copeland <ben.copeland@linaro.org>