.gitlab-ci.yml: add CI job to validate job-definitions
Add a CI job to validate generated LAVA job-definitions files directly
from LAVA. This is needed, as currently in unit tests, the generated
LAVA job definition files are only validated to check whether they are
valid YAML or not but this overlooks the fact that LAVA job definition
files can still be invalid if LAVA don't expects some fields or the
format is not properly aligned with LAVA job schema. This CI job will
run only when there are changes in job definition files, i.e the
location '/test/unit/refs/definitions/*'.

Signed-off-by: Alok Ranjan <alok.ranjan@linaro.org>