fvp-aemva: add cpio/initrd rootfs support
FVP boots via UEFI which requires a block device for the rootfs.
Previously this worked with ext4 image, this change adds support for
cpio rootfs by passing it as an initrd via the EFI startup.nsh script.

For CPIO rootfs we use 'initrd=<filename>' parameter instead of
root=/dev/vda. We also skip the virtioblockdevice.image_path and convert
the URL for LAVA to format corrrectly.

This allows FVP to boot with the same cpio rootfs images used by QEMU,
but without the need for any conversion.

Signed-off-by: Ben Copeland <ben.copeland@linaro.org>